refactor: 修复 Mypy 类型检查问题 (#1445)
fixes #1442

好的，这是将pull request总结翻译成中文的结果：

## Sourcery 总结

重构并修复项目中多个文件的 Mypy 类型检查问题

Bug 修复：
- 解决 Mypy 类型检查警告和错误
- 修复潜在的类型相关的运行时问题
- 添加缺失的类型注解并改进现有类型提示

增强功能：
- 改进各种适配器和核心模块中的类型注解和类型安全
- 添加类型提示并修复 Mypy 检测到的类型相关问题
- 增强整个项目的类型检查合规性

<details>
<summary>Original summary in English</summary>

## Summary by Sourcery

Refactor and fix Mypy type checking issues across multiple files in the
project

Bug Fixes:
- Resolve Mypy type checking warnings and errors
- Fix potential type-related runtime issues
- Add missing type annotations and improve existing type hints

Enhancements:
- Improve type annotations and type safety in various adapters and core
modules
- Add type hints and fix type-related issues detected by Mypy
- Enhance type checking compliance across the project

</details>

+++ kirara_ai/workflow/core/dispatch/rules/system_rules.py
@@ -19,7 +19,7 @@ class RandomChanceMatchRule(DispatchRule):
     config_class = RandomChanceRuleConfig
     type_name = "random"
 
-    def __init__(self, chance: float, workflow_registry: WorkflowRegistry, workflow_id: str):
+    def __init__(self, chance: int, workflow_registry: WorkflowRegistry, workflow_id: str):
         super().__init__(workflow_registry, workflow_id)
         self.chance = chance
